+++ gammapy/irf/tests/test_effective_area.py
@@ -7,6 +7,8 @@ from astropy.units import Quantity
 from astropy.coordinates import Angle
 from astropy.utils.data import get_pkg_data_filename
 from astropy.tests.helper import pytest
+from astropy.io import fits
+
 from ...irf import EffectiveAreaTable2D, EffectiveAreaTable, abramowski_effective_area
 from ...datasets import load_arf_fits_table, load_aeff2D_fits_table
 
@@ -41,24 +43,21 @@ def test_EffectiveAreaTable():
     assert arf.info() == info_str
 
 
-def test_EffectiveAreaTable_write():
-    from tempfile import NamedTemporaryFile
-    from astropy.io import fits
-
+def test_EffectiveAreaTable_write(tmpdir):
     # Read test psf file
     psf = EffectiveAreaTable.from_fits(load_arf_fits_table())
 
+    filename = str(tmpdir.join('effarea_test.fits'))
     # Write it back to disk
-    with NamedTemporaryFile(suffix='.fits') as psf_file:
-        psf.write(psf_file.name)
+    psf.write(filename)
 
-        # Verify checksum
-        hdu_list = fits.open(psf_file.name)
-        # TODO: replace this assert with something else.
-        # For unknown reasons this verify_checksum fails non-deterministically
-        # see e.g. https://travis-ci.org/gammapy/gammapy/jobs/31056341#L1162
-        # assert hdu_list[1].verify_checksum() == 1
-        assert len(hdu_list) == 2
+    # Verify checksum
+    hdu_list = fits.open(filename)
+    # TODO: replace this assert with something else.
+    # For unknown reasons this verify_checksum fails non-deterministically
+    # see e.g. https://travis-ci.org/gammapy/gammapy/jobs/31056341#L1162
+    # assert hdu_list[1].verify_checksum() == 1
+    assert len(hdu_list) == 2
 
 
 INTERPOLATION_METHODS = ['linear', 'spline']